Moving git functions to their own package
 - I don't want to make people that aren't interested in this functionality download the dependencies.
